已完成：用户签到功能 回复功能 未完成：帖子功能 Controller层

+++ src/main/java/com/eric/controller/ResponseController.java
@@ -33,7 +33,7 @@ public class ResponseController {
     }
 
     @RequestMapping("/{commentid}")
-    public List<ResponseVo> getResponseByComId(@RequestParam("commentid") int commentid){
+    public List<ResponseVo> getResponseByComId(@RequestParam("commentid") Long commentid){
         return responseService.queryAllByCommentId(commentid);
     }
     @RequestMapping("/delete/{responseid}")
